[wip] add REST API endpoint to generate GeoJSON from a filter
not completely functional yet
